requirements.txt
django==3.0.4
dj-database-url==0.5.0
django-registration==3.1
djangorestframework==3.11.0
drf-yasg==1.17.1
requests==2.23.0
gunicorn==20.0.4 
psycopg2-binary==2.8.4 
whitenoise==5.0.1
sendgrid-django==4.2.0